fix: Fix empty cell value for date\dateTime columns with custom format (#129)
When `applyFormatOnInit` is true, the `applyFormat` method is called twice on first build. If the date\dateTime column format is not accepted by `DateTime.parse`, the second call to `applyFormat` will result in an empty cell value due to the `null` result from `DateTime.tryParse`.

This commit improve the following methods `applyFormat`, `isValid` and `makeCompareValue` to be more compatible and handles values of type String or DateTime.